Polish: fmt/clippy clean, add license/metadata/CI, drop dead demo
- Apply rustfmt and fix all clippy warnings (unnecessary casts, collapsible
  ifs, redundant closure).
- Add MIT LICENSE and fill in Cargo.toml metadata (description, license,
  repository, readme, keywords, categories).
- Add a GitHub Actions quality workflow (fmt --check, clippy -D warnings,
  test, release build) to enforce the pre-commit gate.
- Remove the leftover `tap` demo subcommand/module (superseded by the engine).
- Refresh stale crate/config module docs (drop "Milestone 1 PoC"; note
  matching is by vendor/product id, not registry_id).

+++ src/hid.rs
@@ -25,11 +25,11 @@ pub struct MouseDevice {
 
 /// Enumerate all HID devices and keep the ones that look like pointers/mice.
 pub fn list_mice() -> Vec<MouseDevice> {
-    let manager = IOHIDManager::new(None, kIOHIDOptionsTypeNone as u32);
+    let manager = IOHIDManager::new(None, kIOHIDOptionsTypeNone);
 
     // Match everything (None), then filter by usage below. Avoids building a match dictionary.
     unsafe { manager.set_device_matching(None) };
-    let _ = manager.open(kIOHIDOptionsTypeNone as u32);
+    let _ = manager.open(kIOHIDOptionsTypeNone);
 
     let Some(set) = manager.devices() else {
         return Vec::new();
@@ -40,8 +40,7 @@ pub fn list_mice() -> Vec<MouseDevice> {
         .map(read_device)
         // Generic Desktop (0x01), Mouse (0x02) or Pointer (0x01).
         .filter(|d| {
-            d.primary_usage_page == Some(0x01)
-                && matches!(d.primary_usage, Some(0x02) | Some(0x01))
+            d.primary_usage_page == Some(0x01) && matches!(d.primary_usage, Some(0x02) | Some(0x01))
         })
         // Never surface the built-in Apple trackpad/keyboard: remapping the internal pointer is a
         // great way to lock yourself out, so it's excluded from discovery entirely.
